requirements.txt
absl-py==1.4.0
addict==2.4.0
albumentations==1.3.0
asttokens==2.2.1
backcall==0.2.0
basicsr==1.4.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
efficientnet-pytorch==0.7.1
einops==0.6.1
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
fsspec==2023.6.0
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.12.2
jedi==0.19.0
joblib==1.2.0
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
munch==4.0.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
protobuf==4.24.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==1.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rpe-index==1.2.0
rsa==4.9
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
six==1.16.0
stack-data==0.6.2
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.2
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.65.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
yapf==0.40.1
zipp==3.16.2
